Updated the authors, description, name and version.
For the authors field, I have added @asiekierka to to the authors. In the description I added a period, because why not? And in the name field I change Buildcraft to BuildCraft as the C in BuildCraft is capitalized. I also updated the Version field to 6.3.4 as that's the latest stable version of BuildCraft.