Update package.json to properly specify deps
Update the `package.json` file to properly specify the dependency on
`crafting-guide-common` by referring to it's GitHub repository. This
should allow someone working on this package to have NPM donwload it
like anything else (including CircleCI).

Also change the `link-deps` Grunt command to `use-local-deps` and
make it remove the NPM-downloaded version and replace it with a
symlink to a local copy.