Include Twitter script after Twitter button
The Twitter button would frequently fail to render properly because
the script which dynamically creates the proper widget would load
before the DOM element which serves as the placeholder for the widget
hadn't yet been created.  This change ensures that the widget script
is always loaded after the placeholder has been created.