Consider item craftable if multiblock
Because the item page use the `craftable` attribute of the item to
decide whether to show the "Show Crafting Plan" button, it was not
showing that button for multiblocks (which don't have any normal
recipes). This change make it so that an item is also considered to
be craftable if it has a multiblock recipe.

fixes #256